# packages/contract/generated/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Clone Instagram API",
    "version": "0.1.0",
...
  },
  "servers": [
    {
      "url": "/",
      "description": "Même origine ; l'hôte dépend de l'environnement (local ou démo, ADR-009)."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/media/uploads": {
      "post": {
        "operationId": "requestMediaUpload",
        "tags": [
          "media"
        ],
        "summary": "Demander une intention d'upload",
        "description": "Crée le média en `pending_upload` et renvoie une URL d'upload présignée (ADR-008).\nLe fichier est envoyé directement à `uploadUrl` par un `PUT` du contenu brut, avec le `Content-Type` déclaré, avant `expiresAt` ;\nl'API ne transporte jamais le fichier. En T3 : image JPEG ou PNG de 20 Mo au plus, pour une photo de profil.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "kind",
                  "purpose",
                  "mimeType",
                  "sizeBytes"
                ],
                "properties": {
                  "kind": {
                    "type": "string",
                    "description": "`video` arrive avec les reels (P1).",
                    "enum": [
                      "image"
                    ]
                  },
                  "purpose": {
                    "type": "string",
                    "description": "`post` arrive en T6a.",
                    "enum": [
                      "avatar"
                    ]
                  },
                  "mimeType": {
                    "type": "string",
                    "enum": [
                      "image/jpeg",
                      "image/png"
                    ]
                  },
                  "sizeBytes": {
                    "type": "integer",
                    "description": "Taille déclarée du fichier ; le worker revérifie le fichier réel.",
                    "minimum": 1,
                    "maximum": 20971520
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Intention d'upload créée.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "mediaId",
                    "uploadUrl",
                    "expiresAt"
                  ],
                  "properties": {
                    "mediaId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "uploadUrl": {
                      "type": "string",
                      "format": "uri",
                      "description": "URL présignée ; `PUT` du contenu brut du fichier."
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Entrée non conforme au schéma (`validation_failed`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "description": "Erreur au format Problem Details (RFC 9457).",
                  "additionalProperties": false,
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "code"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "Identifiant du type de problème (`about:blank` par défaut).",
                      "examples": [
                        "about:blank"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "description": "Résumé lisible du statut HTTP.",
                      "examples": [
                        "Not Found"
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "examples": [
                        404
                      ]
                    },
                    "detail": {
                      "type": "string",
                      "description": "Explication lisible de cette occurrence."
                    },
                    "code": {
                      "type": "string",
                      "description": "Code stable, destiné au client (ex. `validation_failed`, `username_taken`).",
                      "examples": [
                        "route_not_found"
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "JWT absent ou invalide (`unauthenticated`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "description": "Erreur au format Problem Details (RFC 9457).",
                  "additionalProperties": false,
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "code"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "Identifiant du type de problème (`about:blank` par défaut).",
                      "examples": [
                        "about:blank"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "description": "Résumé lisible du statut HTTP.",
                      "examples": [
                        "Not Found"
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "examples": [
                        404
                      ]
                    },
                    "detail": {
                      "type": "string",
                      "description": "Explication lisible de cette occurrence."
                    },
                    "code": {
                      "type": "string",
                      "description": "Code stable, destiné au client (ex. `validation_failed`, `username_taken`).",
                      "examples": [
                        "route_not_found"
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Onboarding non terminé (`profile_not_found`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "description": "Erreur au format Problem Details (RFC 9457).",
                  "additionalProperties": false,
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "code"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "Identifiant du type de problème (`about:blank` par défaut).",
                      "examples": [
                        "about:blank"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "description": "Résumé lisible du statut HTTP.",
                      "examples": [
                        "Not Found"
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "examples": [
                        404
                      ]
                    },
                    "detail": {
                      "type": "string",
                      "description": "Explication lisible de cette occurrence."
                    },
                    "code": {
                      "type": "string",
                      "description": "Code stable, destiné au client (ex. `validation_failed`, `username_taken`).",
                      "examples": [
                        "route_not_found"
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erreur inattendue (`internal_error`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "description": "Erreur au format Problem Details (RFC 9457).",
                  "additionalProperties": false,
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "code"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "Identifiant du type de problème (`about:blank` par défaut).",
                      "examples": [
                        "about:blank"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "description": "Résumé lisible du statut HTTP.",
                      "examples": [
                        "Not Found"
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "examples": [
                        404
                      ]
                    },
                    "detail": {
                      "type": "string",
                      "description": "Explication lisible de cette occurrence."
                    },
                    "code": {
                      "type": "string",
                      "description": "Code stable, destiné au client (ex. `validation_failed`, `username_taken`).",
                      "examples": [
                        "route_not_found"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/media/{id}/complete": {
      "post": {
        "operationId": "completeMediaUpload",
        "tags": [
          "media"
        ],
        "summary": "Confirmer l'envoi du fichier",
        "description": "Passe le média de `pending_upload` à `uploaded` et lance son traitement. Rappelé sur un média déjà `uploaded`, relance le traitement sans doublon.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Média en attente de traitement.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Un de mes médias.",
                  "additionalProperties": false,
                  "required": [
                    "id",
                    "kind",
                    "purpose",
                    "status"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "kind": {
                      "type": "string",
                      "description": "`video` arrive avec les reels (P1).",
                      "enum": [
                        "image"
                      ]
                    },
                    "purpose": {
                      "type": "string",
                      "description": "`post` arrive en T6a.",
                      "enum": [
                        "avatar"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "description": "`pending_upload` → `uploaded` → `processing` → `ready` | `failed` (ADR-008).",
                      "enum": [
                        "pending_upload",
                        "uploaded",
                        "processing",
                        "ready",
                        "failed"
                      ]
                    },
                    "variants": {
                      "description": "Présentes quand le média est `ready`.",
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "thumb",
                        "medium",
                        "large"
                      ],
                      "properties": {
                        "thumb": {
                          "type": "string",
                          "format": "uri",
                          "description": "150 px de large."
                        },
                        "medium": {
                          "type": "string",
                          "format": "uri",
                          "description": "640 px de large."
                        },
                        "large": {
                          "type": "string",
                          "format": "uri",
                          "description": "1080 px de large."
                        }
                      }
                    },
                    "failureReason": {
                      "description": "Présent quand le média est `failed`.",
                      "type": "string",
                      "enum": [
                        "invalid_image",
                        "file_too_large",
                        "processing_error"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Entrée non conforme au schéma (`validation_failed`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "description": "Erreur au format Problem Details (RFC 9457).",
                  "additionalProperties": false,
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "code"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "Identifiant du type de problème (`about:blank` par défaut).",
                      "examples": [
                        "about:blank"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "description": "Résumé lisible du statut HTTP.",
                      "examples": [
                        "Not Found"
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "examples": [
                        404
                      ]
                    },
                    "detail": {
                      "type": "string",
                      "description": "Explication lisible de cette occurrence."
                    },
                    "code": {
                      "type": "string",
                      "description": "Code stable, destiné au client (ex. `validation_failed`, `username_taken`).",
                      "examples": [
                        "route_not_found"
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "JWT absent ou invalide (`unauthenticated`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "description": "Erreur au format Problem Details (RFC 9457).",
                  "additionalProperties": false,
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "code"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "Identifiant du type de problème (`about:blank` par défaut).",
                      "examples": [
                        "about:blank"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "description": "Résumé lisible du statut HTTP.",
                      "examples": [
                        "Not Found"
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "examples": [
                        404
                      ]
                    },
                    "detail": {
                      "type": "string",
                      "description": "Explication lisible de cette occurrence."
                    },
                    "code": {
                      "type": "string",
                      "description": "Code stable, destiné au client (ex. `validation_failed`, `username_taken`).",
                      "examples": [
                        "route_not_found"
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Média inexistant ou d'un autre utilisateur (`media_not_found`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "description": "Erreur au format Problem Details (RFC 9457).",
                  "additionalProperties": false,
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "code"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "Identifiant du type de problème (`about:blank` par défaut).",
                      "examples": [
                        "about:blank"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "description": "Résumé lisible du statut HTTP.",
                      "examples": [
                        "Not Found"
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "examples": [
                        404
                      ]
                    },
                    "detail": {
                      "type": "string",
                      "description": "Explication lisible de cette occurrence."
                    },
                    "code": {
                      "type": "string",
                      "description": "Code stable, destiné au client (ex. `validation_failed`, `username_taken`).",
                      "examples": [
                        "route_not_found"
                      ]
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Média déjà en traitement, prêt ou en échec (`media_invalid_transition`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "description": "Erreur au format Problem Details (RFC 9457).",
                  "additionalProperties": false,
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "code"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "Identifiant du type de problème (`about:blank` par défaut).",
                      "examples": [
                        "about:blank"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "description": "Résumé lisible du statut HTTP.",
                      "examples": [
                        "Not Found"
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "examples": [
                        404
                      ]
                    },
                    "detail": {
                      "type": "string",
                      "description": "Explication lisible de cette occurrence."
                    },
                    "code": {
                      "type": "string",
                      "description": "Code stable, destiné au client (ex. `validation_failed`, `username_taken`).",
                      "examples": [
                        "route_not_found"
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erreur inattendue (`internal_error`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "description": "Erreur au format Problem Details (RFC 9457).",
                  "additionalProperties": false,
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "code"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference",
                      "description": "Identifiant du type de problème (`about:blank` par défaut).",
                      "examples": [
                        "about:blank"
                      ]
                    },
                    "title": {
                      "type": "string",
                      "description": "Résumé lisible du statut HTTP.",
                      "examples": [
                        "Not Found"
                      ]
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 100,
                      "maximum": 599,
                      "examples": [
                        404
                      ]
                    },
                    "detail": {
                      "type": "string",
                      "description": "Explication lisible de cette occurrence."
                    },
                    "code": {
                      "type": "string",
                      "description": "Code stable, destiné au client (ex. `validation_failed`, `username_taken`).",
                      "examples": [
                        "route_not_found"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token Supabase Auth."
      }
    },
...
  }
}